fix: split components
