+++ packages/wallets/src/components/WalletsCarouselContent/WalletsCarouselContent.scss
@@ -1,4 +1,4 @@
-.wallets-carousel {
+.wallets-carousel-content {
     background-color: #f2f3f4;
     width: 100vw;
     padding: 2rem;
