+++ packages/wallets/src/components/WalletsCarouselContent/index.ts
@@ -1,0 +1,1 @@
+export { default as WalletsCarouselContent } from './WalletsCarouselContent';
